+++ src/pages/helpers/task/HelperTaskPage.tsx
@@ -12,12 +12,12 @@ import HelperTaskView from "./HelperTaskView";
 const HelperTaskPage: React.FC = () => {
   const { id } = useParams();
   const task = usePromise(
-    (signal?: AbortSignal): Promise<HelperTask> => {
+    async (signal?: AbortSignal) => {
       const task_id = parseInt(id ?? "NaN");
       if (isNaN(task_id)) {
         throw new Error("Invalid task ID");
       } else {
-        return client.helpers.getTaskById(task_id, signal);
+        return await client.helpers.getTaskById(task_id, signal);
       }
     },
     [id],