+++ src/pages/home/HomePage.tsx
@@ -2,10 +2,12 @@ import ReadingBox from "@/components/layout/ReadingBox";
 
 import MyTasksView from "./MyTasksView";
 import ProfileView from "./ProfileView";
+import HelpersSpeedDial from "../helpers/components/HelpersSpeedDial";
 
 const HomePage: React.FC = () => {
   return (
     <ReadingBox>
+      <HelpersSpeedDial />
       <MyTasksView />
       <ProfileView />
     </ReadingBox>