+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Share via Email and WhatsApp buttons
 - Extra confirmation checks when updating helper tasks
 - Permissions page for admins
+- Audit log page for admins
 - Prefetch shared data for admins and editors
 
 ### Changed